sanitize names in comments for code generated Functions and MXFunction calls
a31a680 made code generation also print function names (in comments). This was dangerous
because users could escape the comments and write arbitrary c code. This commit adds a
method Function/FunctionInternal::getSanitizedNames() which returns a safe version
of the name with non-alphanumeric characters replaced with '_'. This method is used
for the code generation.