#1106: add WITH_SNOPT_FORCE_DUMMY option
This commit closes #1106 and reverts the behavior to the way it was before I
started tinkering for #1106. The user now uses SNOPT=their_snopt and FIND_PACKAGE(SNOPT)
tries to compile with it.

The difference now is that WITH_SNOPT_FORCE_DUMMY (which is off by default) will
ignore the libraries found by FIND_PACKAGE and instead link with a generic -lsnopt7.
We also have to build this empty dummy library, but we don't install it.

There are no more weak symbols