Uniform FooToBar semantics for lyapunov solvers
As per private discussion with Joel.

Consider QpToNlp

Even though we are actually transforming from the interface of a 'NLP
solver' to look and feel like a 'QP solver',
we adopt the point of view that we are transforming the 'QP problem'
into an 'NLP problem'.
This feels more like linux 'pdf2ps' and extends when we want to do the
transformation backwards. I.e. recognising that an NLP is in fact a QP
and converting/treating it as such. It makes sense to call such
transformation 'NlpToQp'.