Added a cache for all UnarySX and BinarySX objects.
This reduces the number of nodes in an expression by avoiding duplicate
nodes which speeds up the evaluation.
(My test case shows a reduction of the number of nodes by a factor of
 more than 2)

This commit may be related to issue #1540.

Note that the implementation of the removal of nodes from the cache
(when the node is deleted) is quite complicated. This is only required
due to the "rather complex destructor" ~BinarySX to avoid a recursive
call of this destructor.