Optimize get_nz by avoiding integer division
Integer divisions are very costly, and typically take ~20-26 clocks
cycles for 32-bit integers, and ~85-100 clock cycles for 64-bit
integers. The default size of indices was changed to 64-bit in commit
a0c6de6c6, causing a significant performance hit in one of the get_nz
methods.

This commit rewrites said get_nz method to avoid divisions, and instead
use additions. These have a typical latency of 1 cycle, and a throughput
of 0.25-0.33 cycles, regardless of argument size.