Make linspace() non-recursive
For a "linspace(a, b, num)" call with "a" and/or "b" symbolic expressions,
the recursive implementation would lead to a graph of "num" depth. For
large "num" this can lead to e.g. a very slow "is_equal" check due to the
depth requirement.

Another reason to prefer a multiplication over recursive addition is
eliminating the compounding error due to rounding when calling linspace
with numeric arguments.