CbcInterface: Fix erroneous passing of hot_start as cmd argument
Hot start is already handled separately by enabling the call to
setBestSolution, and should therefore not be passed explicitly to CBC as
well. Up until this commit, it was passed as a (virtual) command line
argument, resulting in CBC printing a warning that it does not recognize
an option call "hot_start" (and its argument).