CbcInterface: Fix disappearing log messages
Messages from Cbc were disappearing in the message handler. We can
remove it entirely, which makes sure the message show up again in
stdout/stderr. Note that unlike with the Clp interface, the user can
properly capture stdout and stderr as-is. No special capturing/piping of
messages is therefore necessary.